--- app/src/main/res/drawable/save.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="800dp"
-    android:height="800dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-  <path
-      android:pathData="M6.75,6L7.5,5.25H16.5L17.25,6V19.316L12,16.205L6.75,19.316V6ZM8.25,6.75V16.684L12,14.462L15.75,16.684V6.75H8.25Z"
-      android:fillColor="@color/gray"
-      android:fillType="evenOdd"/>
-</vector>
