--- app/src/main/res/drawable/shapeemail.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-    <solid
-        android:color="@android:color/transparent">
-    </solid>
-
-    <corners
-        android:radius="25dp">
-    </corners>
-
-    <stroke
-        android:width="1dp"
-        android:color="#5B4472">
-    </stroke>
-
-    <padding
-        android:top="0dp"
-        android:right="10dp"
-        android:left="10dp"
-        android:bottom="0dp">
-    </padding>
-
-</shape>
