--- app/src/main/res/drawable/side_nav_bar.xml
@@ -1,9 +1,0 @@
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <gradient
-        android:type="linear"
-        android:startColor="#9A81CB"
-        android:endColor="#C8C1D4"
-        android:angle="0"
-        />
-</shape>
