+++ app/src/main/res/layout/activity_news_show.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.NewsShowActivity">
+    tools:context=".EnewspaperActivity$NewsShowActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
