+++ app/src/main/res/layout/fragment_notification.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/homefragment"
-    tools:context=".ui.home.HomeFragment">
+    tools:context=".EnewspaperActivity$HomeFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerView"
