+++ app/src/main/res/layout/fragment_search.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/eventfragment"
-    tools:context=".ui.gallery.SearchFragment"
+    tools:context=".EnewspaperActivity$SearchFragment"
     android:orientation="vertical">
